Added specific_entropy, reorganized some stuff
